learn and test the experiment for mrp
dlpmln:
In the learn function, I modify the part of neural network backward, previously we fix the reshape size to be 10 which is not general.

In the testConstraint part, I add one more argument opt, so that we can also test optimal constraint. the way that I implement is not general enough, since I need to know the nn atom for each neural network, for example, in the mrp experiment, I need to know that neural network “m” is relative to nn_edge(), given the name of nn_edge so that I can find this atom in the optimal models given by Clingo. And then I can check if the prediction of neural network “m” is one of the optimal solution. Also I need to know the term that map to each nn atom, for example in the same experiment, the term maps to neural network “m” is “g”. This info can also be got when we parse the program.